requirements.txt
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.24.1
marshmallow==3.10.0
SQLAlchemy==1.3.22
Flask-Migrate==2.5.3
Werkzeug==1.0.1
alembic==1.4.3
mysqlclient==2.0.3
gunicorn==20.0.4
SQLAlchemy-Utils==0.36.8
Flask-RESTful==0.3.8
webargs==7.0.1
passlib==1.7.4
bcrypt==3.2.0
Flask-SeaSurf==0.3.0
PyJWT==2.0.1
redis==3.5.3
hiredis==1.1.0
